added additional if condition to avoid over-alignment when doing the try_calibrate()

+++ arduino_motor/mains/Serial.ino
@@ -13,7 +13,7 @@ bool SerialCommnder::receive_exec_command() {
   
   if(Serial.available()>0) {
     int function_code;
-    double parameter;
+    long parameter;
 
     long comamnd_int = Serial.parseInt();
     function_code = comamnd_int/100000;
